requirements.txt
Flask==2.3.3
streamlit==1.28.0
openai==1.12.0
python-dotenv==1.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
